fix filter di bagian productInDeposit dan memperbarui bagian layout detail deposit,refactoring di beberapa code

+++ app/src/main/java/com/example/konsinyasiapp/entities/ProductInDeposit.kt
@@ -21,6 +21,6 @@ class ProductInDeposit(
     var jumlahQuantity: String,
 
     @ColumnInfo(name = "id_deposit")
-    var idDeposit: Int,
+    var idDeposit: Long,
 
 ) : Parcelable
